requirements.txt
fastapi[standard]
sqlalchemy
fastapi
pydantic
uvicorn
uuid7
httpx
psycopg2-binary
python-jose[cryptography]
passlib
python-multipart
slowapi
pycountry
python-dotenv